refactor: Fix HadesUserModel
