+++ common/src/main/java/com/wynntils/core/components/Managers.java
@@ -18,7 +18,7 @@ import com.wynntils.core.net.NetManager;
 import com.wynntils.core.net.UrlManager;
 import com.wynntils.core.net.athena.UpdateManager;
 import com.wynntils.core.net.athena.WynntilsAccountManager;
-import com.wynntils.core.net.hades.model.HadesManager;
+import com.wynntils.core.net.hades.HadesManager;
 import com.wynntils.core.net.translation.TranslationManager;
 import com.wynntils.core.notifications.NotificationManager;
 import com.wynntils.core.splashes.SplashManager;
