+++ common/src/main/java/com/wynntils/core/net/hades/HadesManager.java
@@ -2,15 +2,15 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.core.net.hades.model;
+package com.wynntils.core.net.hades;
 
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Manager;
 import com.wynntils.core.components.Managers;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.net.athena.WynntilsAccountManager;
-import com.wynntils.core.net.hades.HadesClientHandler;
 import com.wynntils.core.net.hades.event.HadesEvent;
+import com.wynntils.core.net.hades.objects.HadesUser;
 import com.wynntils.core.net.hades.objects.PlayerStatus;
 import com.wynntils.features.user.HadesFeature;
 import com.wynntils.hades.objects.HadesConnection;
@@ -34,6 +34,7 @@ import java.util.List;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Stream;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.network.chat.ClickEvent;
@@ -48,6 +49,7 @@ public final class HadesManager extends Manager {
     private static final int MS_PER_PING = 1000;
 
     private HadesConnection hadesConnection;
+    private HadesUserRegistry userRegistry = new HadesUserRegistry();
     private int tickCountUntilUpdate = 0;
     private PlayerStatus lastSentStatus;
     private ScheduledExecutorService pingScheduler;
@@ -60,6 +62,10 @@ public final class HadesManager extends Manager {
         }
 
         Managers.WynntilsAccount.onLoginRun(this::onLogin);
+    }
+
+    public Stream<HadesUser> getHadesUsers() {
+        return userRegistry.getHadesUserMap().values().stream();
     }
 
     private void onLogin() {
@@ -75,7 +81,7 @@ public final class HadesManager extends Manager {
                     .setAddress(InetAddress.getByName("io.wynntils.com"), 9000)
                     .setDirection(PacketDirection.SERVER)
                     .setCompressionThreshold(256)
-                    .setHandlerFactory(HadesClientHandler::new)
+                    .setHandlerFactory(a -> new HadesClientHandler(a, userRegistry))
                     .buildClient();
 
             tickCountUntilUpdate = 0;
@@ -132,6 +138,7 @@ public final class HadesManager extends Manager {
 
     @SubscribeEvent
     public void onWorldStateChange(WorldStateEvent event) {
+        userRegistry.reset();
         if (event.isFirstJoinWorld()) {
             if (!isConnected()) {
                 MutableComponent failed = Component.literal("Welps! Trying to connect to Hades failed.")
@@ -218,6 +225,10 @@ public final class HadesManager extends Manager {
         if (!isConnected()) return;
 
         hadesConnection.sendPacketAndFlush(new HCPacketSocialUpdate(List.of(), PacketAction.RESET, socialType));
+    }
+
+    public void resetHadesUsers() {
+        userRegistry.getHadesUserMap().clear();
     }
 
     private boolean isConnected() {
